// frontend/src/app/api/model/room.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Room { 
    /**
     * The unique identifier of this room
     */
    readonly id?: number;
    /**
     * The friendly description of this room
     */
    description: string | null;
    /**
     * The room number according to the Maisel
     */
    roomNumber: number;
    /**
     * The main vlan number assigned to this room
     */
    vlan: number;
}